docs: fix README package names (#16)
* docs: fix README package names

As of #3, the cli package has been called 'nilla-cli' in `nilla.nix`,
however the README incorrectly still calls it 'nilla'

* fix!: correct package name in flake

After #3, the naming convention for packages changed such that
everywhere else our package is called 'nilla-cli'. Unfortunately, we
forgot to update the flake at the same time.

No docs update is needed for this change, as the docs already refer to
the package by the correct name